Late Night Motivation :>
-Added the name of each item under the crosshair when hovering (Only works for things with 'ItemData')

-Changes crosshair color to red when hovering over an interactable item, ie. light switch & Cabinets (WORKS ON COMPUTER SCREEN TOO)

-fixed player stamina starting at half

-EMF now has new light that indicates whether the tool is on or off

-[Attempted] Items are not used when looking at interactables and clicking (commented out bc does not wanna work :<)

-ITEMS BREAK WHEN NO MORE DURABILITY (I had to do a lot of stuff in a lot of files so pls no conflicts...)
 - Also made the crucifix not spammable

pls dont report me...

+++ Assets/Scripts/Inventory Mechanic/ItemData.cs
@@ -18,7 +18,7 @@ public class ItemData : ScriptableObject
     public bool inInventory = false;
     public bool isEquiped;
     public bool isOn;
-    public float durability; // flashlight and the pills but not emf
+    public int durability; // flashlight and the pills but not emf
     public int itemCost;
 
     public float[] cooldownMaxes; // Feel free to designate this arbitrarily per item
